Exception handlers

+++ src/main/java/com/devsuperior/dscommerce/services/ProductService.java
@@ -3,14 +3,20 @@ package com.devsuperior.dscommerce.services;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.devsuperior.dscommerce.dtos.ProductDto;
 import com.devsuperior.dscommerce.entities.Product;
 import com.devsuperior.dscommerce.repositories.ProductRepository;
+import com.devsuperior.dscommerce.services.exceptions.DatabaseException;
+import com.devsuperior.dscommerce.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Component
 public class ProductService {
@@ -22,7 +28,7 @@ public class ProductService {
 	@Transactional(readOnly = true)
 	public ProductDto findById(Long id) {
 		Optional<Product> result = productRepository.findById(id);
-		Product product = result.get();
+		Product product = result.orElseThrow(() -> new ResourceNotFoundException("Id not found"));
 		ProductDto dto = new ProductDto(product);
 		return dto;
 	}
@@ -36,28 +42,39 @@ public class ProductService {
 
 	@Transactional
 	public ProductDto insert(ProductDto dto) {
-		
-		Product entity = new Product();
-		copyDtoToEntity(dto, entity);
-		productRepository.save(entity);
-		
-		return new ProductDto(entity);
+
+			Product entity = new Product();
+			copyDtoToEntity(dto, entity);
+			productRepository.save(entity);
+			return new ProductDto(entity);
+
 	}
 	
 	
 	@Transactional
 	public ProductDto update(Long id, ProductDto dto) {
-		
-		Product entity = productRepository.getReferenceById(id);
-		copyDtoToEntity(dto, entity);
-		entity = productRepository.save(entity);
-		return new ProductDto(entity);
+		try {
+			Product entity = productRepository.getReferenceById(id);
+			copyDtoToEntity(dto, entity);
+			entity = productRepository.save(entity);
+			return new ProductDto(entity);
+		}catch(EntityNotFoundException e) {
+			throw new ResourceNotFoundException(" Product not found ");
+		}
 	}
 	
 	
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete (Long id) {
-		productRepository.deleteById(id);
+		
+		if (!productRepository.existsById(id)) {
+			throw new ResourceNotFoundException("Product not found");
+		}
+		try {
+			productRepository.deleteById(id);    		
+		}catch(DataIntegrityViolationException e ) {
+			throw new DatabaseException("Referential integrity constraint violation ");
+		}
 	}
 	
 	private void copyDtoToEntity(ProductDto dto, Product entity) {
